Fix useful compiler warnings on OSX

+++ userspace/sysdig/chisel.cpp
@@ -53,7 +53,7 @@ void replace_in_place(string &s, const string &search, const string &replace)
 
 void replace_in_place(string& str, string& substr_to_replace, string& new_substr) 
 {
-	uint32_t index = 0;
+	size_t index = 0;
 	uint32_t nsize = new_substr.size();
 
 	while (true)
